refactor: enhance WebGL resource handling for integer attributes
- Updated WebGLResourceRepository to use gl.vertexAttribIPointer for integer attributes, improving type handling for non-floating point data.
- Added default integer values for attributes not present in the primitive to prevent type mismatches with integer shader inputs.
- Changed shader input type for a_joint from vec4 to uvec4 in vertexInOut.glsl and CommonShaderPart.ts for consistency with updated attribute handling.

+++ src/webgl/shaderity_shaders/common/vertexInOut.glsl
@@ -5,11 +5,11 @@ in vec4 a_position;
   const vec4 a_color = vec4(1.0, 1.0, 1.0, 1.0);
 #endif
 in vec3 a_normal;
-in uvec4 a_instanceIds; // <semantic=INSTANCE>
+in uvec4 a_instanceIds;
 in vec2 a_texcoord_0;
 in vec2 a_texcoord_1;
 in vec2 a_texcoord_2;
-in vec4 a_joint;
+in uvec4 a_joint;
 in vec4 a_weight;
 in vec4 a_baryCentricCoord;
 out vec4 v_color;
